update image so it uses size_t instead of int for dimensions
git-svn-id: file:///home/roberto/Git/svn@6507 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 2cbd7dabde2b647e636dd58d70a52e9929587314

+++ trunk/xmipp/applications/programs/micrograph_mark/widget_micrograph.cpp
@@ -1357,9 +1357,10 @@ int AutoParticlePicking::automaticallySelectParticles()
             std::vector<int> toKeep;
             for (int i=0; i<YSIZE(features[1]); i++)
             {
+            	//aux should be a matrix1d
                 MultidimArray<double> aux;
                 features[1].getRow(i,aux);
-                Matrix1D<double> trialFeatures(XSIZE(aux));
+                Matrix1D<double> trialFeatures((int)XSIZE(aux));
                 FOR_ALL_ELEMENTS_IN_ARRAY1D(aux)
 					trialFeatures(i) = aux(i);
                 double cost;
